backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1
Werkzeug==3.1.3
pandas==2.2.3
joblib==1.5.1
numpy==1.26.4
APScheduler==3.11.0
pytz==2024.2
scikit-learn==1.7.0
scikit-surprise==1.1.4
gunicorn==21.2.0
pymysql==1.1.1